requirements.txt
torch
torchcam
torchvision
pillow
numpy
matplotlib